WIP: mdev: start following Laurent Bercot's proposed protocol.
The current protocol expects that we'll read "VAR=VALUE\0" pairs for
one hotplug event.
Multiple hotplug events can arrive in the pipe before read() is called,
so we could end up getting a single handle_event() call for multiple
hotplug events.
So, the proposed protocol is to use a second '\0' at the end of each
hotplug event.