Implement mdev -s in terms of handle_event().
This can save a bit of space (~100 bytes) when MDEV_PIPE is selected,
since it uses some of the same code.
If MDEV_PIPE is not selected, it costs ~26 bytes.

The main point is to allow executing $VAR rules from mdev -s, so that
manually running modprobe is not needed.